pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "private-assistant-intent-engine"
version = "0.0.1"
description = "Intent engine for the private assistant ecosystem"
authors = ["stkr22 <stkr22@github.com>"]
readme = "README.md"
license = "MIT License"
include = ["src"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = "^0.12.3"
private-assistant-commons = "^0.1.4"
spacy = "^3.7.5"
paho-mqtt = "^2.1.0"
word2number = "^1.1"
en-core-web-md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl"}
pydantic = "^2.8.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
mypy = "^1.11.0"
pytest = "^8.3.2"
types-pyyaml = "^6.0.12.20240724"

[tool.poetry.scripts]
private-assistant-intent-engine = "private_assistant_intent_engine.main:app"

[[tool.mypy.overrides]]
module = [
    "spotipy.*",
    "word2number"
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]
ignore = ["E203"]
